Moved helper functions to a new utils package
